Extend device_get_property API
Create a new enum so that a caller can specify the
expected type of property they want to read.
This way the bus logic can use that information
to process the underlying data.
For example in DT world all integer properties
are stored in BE format. In order to get constant
results across different platforms we need to convert
its endianness to match the host.